remove old constants file, update import
